chore: add validation and course quering

+++ app/Controllers/Http/CourseController.js
@@ -2,10 +2,9 @@
 
 /** @typedef {import('@adonisjs/framework/src/Request')} Request */
 /** @typedef {import('@adonisjs/framework/src/Response')} Response */
-/** @typedef {import('@adonisjs/framework/src/View')} View */
 
 const Course = use("App/Models/Course");
-const sharedController = require("./SharedController");
+const { validateAll } = use("Validator");
 
 /**
  * Resourceful controller for interacting with courses
@@ -19,16 +18,44 @@ class CourseController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async index({ auth, response }) {
+  async index({ auth, request, response }) {
     try {
-      const user_id = await auth.user.id;
-      const courses = await Course.query().where("user_id", user_id).fetch();
-      if (!courses) {
-        return response.status(404).send("No couse was found");
+      const { semester, level } = request.all();
+
+      const rules = {
+        semester: "in:1,2",
+        level: "in:100,200,300,400,500,600,700,800,900",
+      };
+
+      const validation = await validateAll(request.all(), rules);
+
+      if (validation.fails()) {
+        return response.status(400).send(validation.messages());
       }
+
+      const user_id = await auth.user.id;
+
+      let courses = [];
+
+      if (level && !semester) {
+        courses = await Course.query()
+          .where({ user_id: user_id, level: level })
+          .fetch();
+      } else if (level && semester) {
+        courses = await Course.query()
+          .where({ user_id: user_id, level: level, semester: semester })
+          .fetch();
+      } else {
+        courses = await Course.query().where("user_id", user_id).fetch();
+      }
+
+      if (courses.rows.length === 0) {
+        return response.status(404).send([]);
+      }
+
       return response.status(200).send(courses);
     } catch (error) {
-      console.log(error)
+      console.log(error);
       return response.status(500).send(error);
     }
   }
@@ -51,18 +78,20 @@ class CourseController {
         semester,
         level,
       } = request.all();
-      const errors = sharedController.missingRequirements(
-        [title, "title"],
-        [grade, "grade"],
-        [code, "code"],
-        [credit_load, "credit_load"],
-        [semester, "semester"],
-        [level, "level"]
-      );
-      if (errors.length) {
-        return response
-          .status(400)
-          .send(`Invalid request. ${errors.join(", ")} are required`);
+
+      const rules = {
+        title: "required",
+        grade: "required|in:A,B,C,D,F",
+        code: "required",
+        credit_load: "required|integer",
+        semester: "required|in:1,2",
+        level: "required|in:100,200,300,400,500,600,700,800,900",
+      };
+
+      const validation = await validateAll(request.all(), rules);
+
+      if (validation.fails()) {
+        return response.status(400).send(validation.messages());
       }
 
       const course = new Course();
@@ -79,6 +108,7 @@ class CourseController {
 
       return response.status(201).send(course);
     } catch (error) {
+      console.log(error);
       return response.status(500).send(error);
     }
   }
@@ -90,11 +120,20 @@ class CourseController {
    * @param {object} ctx
    * @param {Response} ctx.response
    */
-  async show({ params, response }) {
+  async show({ auth, params, response }) {
     try {
       const course = await Course.find(params.id);
+
       if (!course) {
         return response.status(404).send("Course not found");
+      }
+
+      const courseOwner = await course.user().fetch();
+      const requester = await auth.user;
+      if (requester.id !== courseOwner.id) {
+        return response
+          .status(404)
+          .send("You cannot view another user's course");
       }
       return response.status(200).send(course);
     } catch (error) {
@@ -112,18 +151,47 @@ class CourseController {
    */
   async update({ params, request, response }) {
     try {
-      const course = await Course.findOrFail(params.id);
+      const {
+        title,
+        grade,
+        code,
+        credit_load,
+        semester,
+        level,
+      } = request.all();
 
-      course.title = request.body.title || course.title;
-      course.grade = request.body.grade || course.grade;
-      course.code = request.body.code || course.code;
-      course.credit_load = request.body.credit_load || course.credit_load;
-      course.semester = request.body.semester || course.semester;
-      course.level = request.body.level || course.level;
+      const rules = {
+        title: "required",
+        grade: "required|in:A,B,C,D,F",
+        code: "required",
+        credit_load: "required|integer",
+        semester: "required|in:1,2",
+        level: "required|in:100,200,300,400,500,600,700,800,900",
+      };
+
+      const validation = await validateAll(request.all(), rules);
+
+      if (validation.fails()) {
+        return response.status(400).send(validation.messages());
+      }
+
+      const course = await Course.find(params.id);
+
+      if (!course) {
+        return response.status(404).send("Course not found");
+      }
+
+      course.title = title;
+      course.grade = grade;
+      course.code = code;
+      course.credit_load = credit_load;
+      course.semester = semester;
+      course.level = level;
 
       await course.save();
       return response.status(200).send(course);
     } catch (error) {
+      console.log(error);
       return response.status(500).send(error);
     }
   }
@@ -136,9 +204,12 @@ class CourseController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async destroy({ params, request, response }) {
+  async destroy({ params, response }) {
     try {
-      const course = await Course.findOrFail(params.id);
+      const course = await Course.find(params.id);
+      if (!course) {
+        return response.status(404).send("Course not found");
+      }
       await course.delete();
       return response.send("course deleted successfully");
     } catch (error) {
